Reservation Start / End - Final Code cleanup
Modified to support Reservation Start and End functions.
 - Moved the resource power cmds to their own file.
Commented out all unused methods.
basic cleanup for publishing
