+++ Power_Service_Extesnsibility/PowerServiceDriver/deployment.xml
@@ -2,14 +2,14 @@
 <properties>
 
     <!-- The address of the Quali server on which to deploy, mandatory -->
-    <serverRootAddress>WIN-PJH0S8ROU8F</serverRootAddress>
+    <serverRootAddress>localhost</serverRootAddress>
 
     <!-- The port of the Quali server on which to deploy, defaults to "8029" -->
     <port>8029</port>
 
     <!-- The server admin username, password and domain to use when deploying -->
-    <username>admin</username>
-    <password>admin</password>
+    <username>myusername</username>
+    <password>mypassword</password>
     <domain>Global</domain>
 
     <!-- Simple patterns to filter when sending the driver to the server separated by semicolons (e.g. "file.xml;logs/", also supports regular expressions),
