+++ Power_Service_Extesnsibility/PowerServiceDriver/src/debug_info.txt
@@ -1,3 +1,3 @@
-Debugging session started at: 16:10:57 12/08/16
-Python driver process is available for debugging at PID #1356
-Virtual environment is located at `C:\ProgramData\QualiSystems\venv\Power_Service_Extensibility_60_21\Scripts`
+Debugging session started at: 17:11:40 12/19/16
+Python driver process is available for debugging at PID #6352
+Virtual environment is located at `C:\ProgramData\QualiSystems\venv\Power_Service_Extensibility_75_21\Scripts`
