+++ Power_Service_Extesnsibility/PowerServiceDriver/src/driver.py
@@ -40,39 +40,42 @@ class PowerService (ResourceDriverInterface):
         pass
 
 # <editor-fold desc="reservations event handlers">
-    def AfterReservationCreated(self, context, actionDetails, resourcesDetails, serviceDetails):
-        """
-        Intervene after a reservation was created
-        :param ResourceCommandContext context: the context the command runs on
-        :param str actionDetails: details about the action
-        :param str resourcesDetails: details about the resource
-        :param str serviceDetails: details about the service
-        """
-        return ReservationEvents().after_reservation_created(context, actionDetails, resourcesDetails, serviceDetails)
-
-    def AfterReservationExtended(self, context, actionDetails, resourcesDetails, serviceDetails, newEndTime):
-        """
-        Intervene after the end time of a reservation has been changed
-        :param ResourceCommandContext context: the context the command runs on
-        :param str actionDetails: details about the action
-        :param str resourcesDetails: details about the resource
-        :param str serviceDetails: details about the service
-        :param str newEndTime: new end time
-        """
-        return ReservationEvents().after_reservation_extended(context, actionDetails, resourcesDetails, serviceDetails,
-                                                             newEndTime)
-
-    def BeforeReservationExtended(self, context, actionDetails, resourcesDetails, serviceDetails, newEndTime):
-        """
-        Intervene before the end time of a reservation has been changed
-        :param ResourceCommandContext context: the context the command runs on
-        :param str actionDetails: details about the action
-        :param str resourcesDetails: details about the resource
-        :param str serviceDetails: details about the service
-        :param str newEndTime: new end time
-        """
-        return ReservationEvents().before_reservation_extended(context, actionDetails, resourcesDetails, serviceDetails,
-                                                             newEndTime)
+#     def AfterReservationCreated(self, context, actionDetails, resourcesDetails, serviceDetails):
+#         """
+#         Intervene after a reservation was created
+#         :param ResourceCommandContext context: the context the command runs on
+#         :param str actionDetails: details about the action
+#         :param str resourcesDetails: details about the resource
+#         :param str serviceDetails: details about the service
+#         """
+#         # return ReservationEvents().after_reservation_created(context, actionDetails, resourcesDetails, serviceDetails)
+#         pass
+#
+#     def AfterReservationExtended(self, context, actionDetails, resourcesDetails, serviceDetails, newEndTime):
+#         """
+#         Intervene after the end time of a reservation has been changed
+#         :param ResourceCommandContext context: the context the command runs on
+#         :param str actionDetails: details about the action
+#         :param str resourcesDetails: details about the resource
+#         :param str serviceDetails: details about the service
+#         :param str newEndTime: new end time
+#         """
+#         # return ReservationEvents().after_reservation_extended(context, actionDetails, resourcesDetails, serviceDetails,
+#         #                                                      newEndTime)
+#         pass
+#
+#     def BeforeReservationExtended(self, context, actionDetails, resourcesDetails, serviceDetails, newEndTime):
+#         """
+#         Intervene before the end time of a reservation has been changed
+#         :param ResourceCommandContext context: the context the command runs on
+#         :param str actionDetails: details about the action
+#         :param str resourcesDetails: details about the resource
+#         :param str serviceDetails: details about the service
+#         :param str newEndTime: new end time
+#         """
+#         # return ReservationEvents().before_reservation_extended(context, actionDetails, resourcesDetails, serviceDetails,
+#         #                                                      newEndTime)
+#         pass
 
     def BeforeReservationTerminated_Sync(self, context, actionDetails, resourcesDetails, serviceDetails):
         """
@@ -82,7 +85,13 @@ class PowerService (ResourceDriverInterface):
         :param str resourcesDetails: details about the resource
         :param str serviceDetails: details about the service
         """
-        return ReservationEvents().before_reservation_terminated(context, actionDetails, resourcesDetails, serviceDetails)
+        address = context.connectivity.server_address
+        domain = context.reservation.domain
+        token = context.connectivity.admin_auth_token
+        return ReservationEvents(self.get_cs_session(address, domain, token), context.reservation.reservation_id).\
+            before_reservation_terminated(context, actionDetails, resourcesDetails, serviceDetails)
+
+        # return ReservationEvents().before_reservation_terminated(context, actionDetails, resourcesDetails, serviceDetails)
 
     def OnReservationStarted(self, context, resourcesDetails, serviceDetails):
         """
@@ -92,75 +101,80 @@ class PowerService (ResourceDriverInterface):
         :param str serviceDetails: details about the service
         """
 
-        return ReservationEvents().on_reservation_started(context, resourcesDetails, serviceDetails)
+        address = context.connectivity.server_address
+        domain = context.reservation.domain
+        token = context.connectivity.admin_auth_token
 
-    def OnReservationEnded_Sync(self, context, resourcesDetails, serviceDetails):
-        """
-        Intervene before the reservation end date expires (synchronous only)
-        :param ResourceCommandContext context: the context the command runs on
-        :param str resourcesDetails: details about the resource
-        :param str serviceDetails: details about the service
-        """
-        return ReservationEvents().on_reservation_ended(context, resourcesDetails, serviceDetails)
+        return ReservationEvents(self.get_cs_session(address, domain, token), context.reservation.reservation_id).\
+            on_reservation_started(context, resourcesDetails, serviceDetails)
+
+    # def OnReservationEnded_Sync(self, context, resourcesDetails, serviceDetails):
+    #     """
+    #     Intervene before the reservation end date expires (synchronous only)
+    #     :param ResourceCommandContext context: the context the command runs on
+    #     :param str resourcesDetails: details about the resource
+    #     :param str serviceDetails: details about the service
+    #     """
+    #     pass
 
 # </editor-fold>
 
 # <editor-fold desc="connections">
-    def AfterConnectionsChanged(self, context, actionDetails, resourcesDetails, serviceDetails,
-                                addedConnections, removedConnections):
-        """
-        Intervene after a logical connector has been added or removed in a reservation
-        :param ResourceCommandContext context: the context the command runs on
-        :param str actionDetails: details about the action
-        :param str resourcesDetails: details about the resource
-        :param str serviceDetails: details about the service
-        :param str addedConnections: details about the added connections
-        :param str removedConnections: details about the removed connections
-        """
-        return ConnectionEvents().after_connections_changed(context, actionDetails, resourcesDetails, serviceDetails,
-                                                            addedConnections, removedConnections)
-
-    def AfterMyConnectionsChanged(self, context, actionDetails, resourcesDetails, serviceDetails,
-                                  addedConnections, removedConnections):
-        """
-        Intervene after a logical connector has been added or removed to the service to which this driver is associated
-        :param ResourceCommandContext context: the context the command runs on
-        :param str actionDetails: details about the action
-        :param str resourcesDetails: details about the resource
-        :param str serviceDetails: details about the service
-        :param str addedConnections: details about the added connections
-        :param str removedConnections: details about the removed connections
-        """
-        return ConnectionEvents().after_my_connections_changed(context, actionDetails, resourcesDetails, serviceDetails,
-                                                               addedConnections, removedConnections)
-
-    def BeforeConnectionsChanged(self, context, actionDetails, resourcesDetails, serviceDetails,
-                                 addedConnections, removedConnections):
-        """
-        Intervene before a logical connector has been added or removed in a reservation
-        :param ResourceCommandContext context: the context the command runs on
-        :param str actionDetails: details about the action
-        :param str resourcesDetails: details about the resource
-        :param str serviceDetails: details about the service
-        :param str addedConnections: details about the added connections
-        :param str removedConnections: details about the removed connections
-        """
-        return ConnectionEvents().before_connections_changed(context, actionDetails, resourcesDetails, serviceDetails,
-                                                             addedConnections, removedConnections)
-
-    def BeforeMyConnectionsChanged(self, context, actionDetails, resourcesDetails, serviceDetails,
-                                   addedConnections, removedConnections):
-        """
-        Intervene before a logical connector has been added or removed to the service to which this driver is associated
-        :param ResourceCommandContext context: the context the command runs on
-        :param str actionDetails: details about the action
-        :param str resourcesDetails: details about the resource
-        :param str serviceDetails: details about the service
-        :param str addedConnections: details about the added connections
-        :param str removedConnections: details about the removed connections
-        """
-        return ConnectionEvents().before_my_connections_changed(context, actionDetails, resourcesDetails,
-                                                                serviceDetails, addedConnections, removedConnections)
+#     def AfterConnectionsChanged(self, context, actionDetails, resourcesDetails, serviceDetails,
+#                                 addedConnections, removedConnections):
+#         """
+#         Intervene after a logical connector has been added or removed in a reservation
+#         :param ResourceCommandContext context: the context the command runs on
+#         :param str actionDetails: details about the action
+#         :param str resourcesDetails: details about the resource
+#         :param str serviceDetails: details about the service
+#         :param str addedConnections: details about the added connections
+#         :param str removedConnections: details about the removed connections
+#         """
+#         return ConnectionEvents().after_connections_changed(context, actionDetails, resourcesDetails, serviceDetails,
+#                                                             addedConnections, removedConnections)
+#
+#     def AfterMyConnectionsChanged(self, context, actionDetails, resourcesDetails, serviceDetails,
+#                                   addedConnections, removedConnections):
+#         """
+#         Intervene after a logical connector has been added or removed to the service to which this driver is associated
+#         :param ResourceCommandContext context: the context the command runs on
+#         :param str actionDetails: details about the action
+#         :param str resourcesDetails: details about the resource
+#         :param str serviceDetails: details about the service
+#         :param str addedConnections: details about the added connections
+#         :param str removedConnections: details about the removed connections
+#         """
+#         return ConnectionEvents().after_my_connections_changed(context, actionDetails, resourcesDetails, serviceDetails,
+#                                                                addedConnections, removedConnections)
+#
+#     def BeforeConnectionsChanged(self, context, actionDetails, resourcesDetails, serviceDetails,
+#                                  addedConnections, removedConnections):
+#         """
+#         Intervene before a logical connector has been added or removed in a reservation
+#         :param ResourceCommandContext context: the context the command runs on
+#         :param str actionDetails: details about the action
+#         :param str resourcesDetails: details about the resource
+#         :param str serviceDetails: details about the service
+#         :param str addedConnections: details about the added connections
+#         :param str removedConnections: details about the removed connections
+#         """
+#         return ConnectionEvents().before_connections_changed(context, actionDetails, resourcesDetails, serviceDetails,
+#                                                              addedConnections, removedConnections)
+#
+#     def BeforeMyConnectionsChanged(self, context, actionDetails, resourcesDetails, serviceDetails,
+#                                    addedConnections, removedConnections):
+#         """
+#         Intervene before a logical connector has been added or removed to the service to which this driver is associated
+#         :param ResourceCommandContext context: the context the command runs on
+#         :param str actionDetails: details about the action
+#         :param str resourcesDetails: details about the resource
+#         :param str serviceDetails: details about the service
+#         :param str addedConnections: details about the added connections
+#         :param str removedConnections: details about the removed connections
+#         """
+#         return ConnectionEvents().before_my_connections_changed(context, actionDetails, resourcesDetails,
+#                                                                 serviceDetails, addedConnections, removedConnections)
 # </editor-fold>
 
 # <editor-fold desc="resources">
@@ -207,121 +221,121 @@ class PowerService (ResourceDriverInterface):
 # </editor-fold>
 
 # <editor-fold desc="services">
-    def AfterServicesChanged(self, context, actionDetails, resourcesDetails, serviceDetails,
-                             oldServiceAttributeDetails, servicesRemoved, servicesAdded, modifiedServices,
-                             servicesModifiedAttributes):
-        """
-        Intervene after a service is added, removed, or has an attribute changed in reservation
-        :param ResourceCommandContext context: the context the command runs on
-        :param str actionDetails: details about the action
-        :param str resourcesDetails: details about the resource
-        :param str serviceDetails: details about the service
-        :param str oldServiceAttributeDetails: details about the service attributes before the changes
-        :param str servicesRemoved: details about the added services
-        :param str servicesAdded: details about the removed services
-        :param str modifiedServices: details about the modified services
-        :param str servicesModifiedAttributes: details about the service attributes after the changes
-        """
-        address = context.connectivity.server_address
-        domain = context.reservation.domain
-        token = context.connectivity.admin_auth_token
-        return ServicesEvents(self.get_cs_session(address, domain, token), context.reservation.reservation_id).\
-            after_services_changed(context, actionDetails, resourcesDetails, serviceDetails, oldServiceAttributeDetails,
-                                   servicesRemoved, servicesAdded, modifiedServices, servicesModifiedAttributes)
-
-    def BeforeServicesChanged_Sync(self, context, actionDetails, resourcesDetails, serviceDetails,
-                              oldServiceAttributeDetails, servicesRemoved, servicesAdded, modifiedServices,
-                              servicesModifiedAttributes):
-        """
-        Intervene before a service is added, removed, or has an attribute changed in reservation
-        :param ResourceCommandContext context: the context the command runs on
-        :param str actionDetails: details about the action
-        :param str resourcesDetails: details about the resource
-        :param str serviceDetails: details about the service
-        :param str oldServiceAttributeDetails: details about the service attributes before the changes
-        :param str servicesRemoved: details about the added services
-        :param str servicesAdded: details about the removed services
-        :param str modifiedServices: details about the modified services
-        :param str servicesModifiedAttributes: details about the service attributes after the changes
-        """
-        address = context.connectivity.server_address
-        domain = context.reservation.domain
-        token = context.connectivity.admin_auth_token
-        return ServicesEvents(self.get_cs_session(address, domain, token), context.reservation.reservation_id)\
-            .before_services_changed(context, actionDetails, resourcesDetails, serviceDetails,
-                                     oldServiceAttributeDetails, servicesRemoved, servicesAdded, modifiedServices,
-                                     servicesModifiedAttributes)
-
-    def BeforeMyServiceChanged_Sync(self, context, actionDetails, resourcesDetails, serviceDetails,
-                                    oldServiceAttributeDetails, servicesRemoved, servicesAdded, modifiedServices,
-                                    servicesModifiedAttributes):
-        """
-        Intervene before the service with which this driver is associated is added, removed, or has an attribute
-        changed in reservation
-        :param ResourceCommandContext context: the context the command runs on
-        :param str actionDetails: details about the action
-        :param str resourcesDetails: details about the resource
-        :param str serviceDetails: details about the service
-        :param str oldServiceAttributeDetails: details about the service attributes before the changes
-        :param str servicesRemoved: details about the added services
-        :param str servicesAdded: details about the removed services
-        :param str modifiedServices: details about the modified services
-        :param str servicesModifiedAttributes: details about the service attributes after the changes
-        """
-        address = context.connectivity.server_address
-        domain = context.reservation.domain
-        token = context.connectivity.admin_auth_token
-        return ServicesEvents(self.get_cs_session(address, domain, token), context.reservation.reservation_id).\
-            before_my_service_changed(context, actionDetails, resourcesDetails, serviceDetails,
-                                      oldServiceAttributeDetails, servicesRemoved, servicesAdded, modifiedServices,
-                                      servicesModifiedAttributes)
-
-    def BeforeServicesRemoved(self, context, actionDetails, resourcesDetails, serviceDetails,
-                              oldServiceAttributeDetails, servicesRemoved, servicesAdded, modifiedServices,
-                              servicesModifiedAttributes):
-        """
-        Intervene before a service is added, removed, or has an attribute changed in reservation
-        :param ResourceCommandContext context: the context the command runs on
-        :param str actionDetails: details about the action
-        :param str resourcesDetails: details about the resource
-        :param str serviceDetails: details about the service
-        :param str oldServiceAttributeDetails: details about the service attributes before the changes
-        :param str servicesRemoved: details about the added services
-        :param str servicesAdded: details about the removed services
-        :param str modifiedServices: details about the modified services
-        :param str servicesModifiedAttributes: details about the service attributes after the changes
-        """
-        address = context.connectivity.server_address
-        domain = context.reservation.domain
-        token = context.connectivity.admin_auth_token
-        return ServicesEvents(self.get_cs_session(address, domain, token), context.reservation.reservation_id).\
-            before_services_removed(context, actionDetails, resourcesDetails, serviceDetails,
-                                    oldServiceAttributeDetails, servicesRemoved, servicesAdded, modifiedServices,
-                                    servicesModifiedAttributes)
-
-    def BeforeMyServiceRemoved_Sync(self, context, actionDetails, resourcesDetails, serviceDetails,
-                                    oldServiceAttributeDetails, servicesRemoved, servicesAdded, modifiedServices,
-                                    servicesModifiedAttributes):
-        """
-        Intervene before the service with which this driver is associated is added, removed, or has an attribute 
-        changed in reservation
-        :param ResourceCommandContext context: the context the command runs on
-        :param str actionDetails: details about the action
-        :param str resourcesDetails: details about the resource
-        :param str serviceDetails: details about the service
-        :param str oldServiceAttributeDetails: details about the service attributes before the changes
-        :param str servicesRemoved: details about the added services
-        :param str servicesAdded: details about the removed services
-        :param str modifiedServices: details about the modified services
-        :param str servicesModifiedAttributes: details about the service attributes after the changes
-        """
-        address = context.connectivity.server_address
-        domain = context.reservation.domain
-        token = context.connectivity.admin_auth_token
-        return ServicesEvents(self.get_cs_session(address, domain, token), context.reservation.reservation_id).\
-            before_my_service_removed(context, actionDetails, resourcesDetails, serviceDetails,
-                                      oldServiceAttributeDetails, servicesRemoved, servicesAdded, modifiedServices,
-                                      servicesModifiedAttributes)
+#     def AfterServicesChanged(self, context, actionDetails, resourcesDetails, serviceDetails,
+#                              oldServiceAttributeDetails, servicesRemoved, servicesAdded, modifiedServices,
+#                              servicesModifiedAttributes):
+#         """
+#         Intervene after a service is added, removed, or has an attribute changed in reservation
+#         :param ResourceCommandContext context: the context the command runs on
+#         :param str actionDetails: details about the action
+#         :param str resourcesDetails: details about the resource
+#         :param str serviceDetails: details about the service
+#         :param str oldServiceAttributeDetails: details about the service attributes before the changes
+#         :param str servicesRemoved: details about the added services
+#         :param str servicesAdded: details about the removed services
+#         :param str modifiedServices: details about the modified services
+#         :param str servicesModifiedAttributes: details about the service attributes after the changes
+#         """
+#         address = context.connectivity.server_address
+#         domain = context.reservation.domain
+#         token = context.connectivity.admin_auth_token
+#         return ServicesEvents(self.get_cs_session(address, domain, token), context.reservation.reservation_id).\
+#             after_services_changed(context, actionDetails, resourcesDetails, serviceDetails, oldServiceAttributeDetails,
+#                                    servicesRemoved, servicesAdded, modifiedServices, servicesModifiedAttributes)
+#
+#     def BeforeServicesChanged_Sync(self, context, actionDetails, resourcesDetails, serviceDetails,
+#                               oldServiceAttributeDetails, servicesRemoved, servicesAdded, modifiedServices,
+#                               servicesModifiedAttributes):
+#         """
+#         Intervene before a service is added, removed, or has an attribute changed in reservation
+#         :param ResourceCommandContext context: the context the command runs on
+#         :param str actionDetails: details about the action
+#         :param str resourcesDetails: details about the resource
+#         :param str serviceDetails: details about the service
+#         :param str oldServiceAttributeDetails: details about the service attributes before the changes
+#         :param str servicesRemoved: details about the added services
+#         :param str servicesAdded: details about the removed services
+#         :param str modifiedServices: details about the modified services
+#         :param str servicesModifiedAttributes: details about the service attributes after the changes
+#         """
+#         address = context.connectivity.server_address
+#         domain = context.reservation.domain
+#         token = context.connectivity.admin_auth_token
+#         return ServicesEvents(self.get_cs_session(address, domain, token), context.reservation.reservation_id)\
+#             .before_services_changed(context, actionDetails, resourcesDetails, serviceDetails,
+#                                      oldServiceAttributeDetails, servicesRemoved, servicesAdded, modifiedServices,
+#                                      servicesModifiedAttributes)
+#
+#     def BeforeMyServiceChanged_Sync(self, context, actionDetails, resourcesDetails, serviceDetails,
+#                                     oldServiceAttributeDetails, servicesRemoved, servicesAdded, modifiedServices,
+#                                     servicesModifiedAttributes):
+#         """
+#         Intervene before the service with which this driver is associated is added, removed, or has an attribute
+#         changed in reservation
+#         :param ResourceCommandContext context: the context the command runs on
+#         :param str actionDetails: details about the action
+#         :param str resourcesDetails: details about the resource
+#         :param str serviceDetails: details about the service
+#         :param str oldServiceAttributeDetails: details about the service attributes before the changes
+#         :param str servicesRemoved: details about the added services
+#         :param str servicesAdded: details about the removed services
+#         :param str modifiedServices: details about the modified services
+#         :param str servicesModifiedAttributes: details about the service attributes after the changes
+#         """
+#         address = context.connectivity.server_address
+#         domain = context.reservation.domain
+#         token = context.connectivity.admin_auth_token
+#         return ServicesEvents(self.get_cs_session(address, domain, token), context.reservation.reservation_id).\
+#             before_my_service_changed(context, actionDetails, resourcesDetails, serviceDetails,
+#                                       oldServiceAttributeDetails, servicesRemoved, servicesAdded, modifiedServices,
+#                                       servicesModifiedAttributes)
+#
+#     def BeforeServicesRemoved(self, context, actionDetails, resourcesDetails, serviceDetails,
+#                               oldServiceAttributeDetails, servicesRemoved, servicesAdded, modifiedServices,
+#                               servicesModifiedAttributes):
+#         """
+#         Intervene before a service is added, removed, or has an attribute changed in reservation
+#         :param ResourceCommandContext context: the context the command runs on
+#         :param str actionDetails: details about the action
+#         :param str resourcesDetails: details about the resource
+#         :param str serviceDetails: details about the service
+#         :param str oldServiceAttributeDetails: details about the service attributes before the changes
+#         :param str servicesRemoved: details about the added services
+#         :param str servicesAdded: details about the removed services
+#         :param str modifiedServices: details about the modified services
+#         :param str servicesModifiedAttributes: details about the service attributes after the changes
+#         """
+#         address = context.connectivity.server_address
+#         domain = context.reservation.domain
+#         token = context.connectivity.admin_auth_token
+#         return ServicesEvents(self.get_cs_session(address, domain, token), context.reservation.reservation_id).\
+#             before_services_removed(context, actionDetails, resourcesDetails, serviceDetails,
+#                                     oldServiceAttributeDetails, servicesRemoved, servicesAdded, modifiedServices,
+#                                     servicesModifiedAttributes)
+#
+#     def BeforeMyServiceRemoved_Sync(self, context, actionDetails, resourcesDetails, serviceDetails,
+#                                     oldServiceAttributeDetails, servicesRemoved, servicesAdded, modifiedServices,
+#                                     servicesModifiedAttributes):
+#         """
+#         Intervene before the service with which this driver is associated is added, removed, or has an attribute
+#         changed in reservation
+#         :param ResourceCommandContext context: the context the command runs on
+#         :param str actionDetails: details about the action
+#         :param str resourcesDetails: details about the resource
+#         :param str serviceDetails: details about the service
+#         :param str oldServiceAttributeDetails: details about the service attributes before the changes
+#         :param str servicesRemoved: details about the added services
+#         :param str servicesAdded: details about the removed services
+#         :param str modifiedServices: details about the modified services
+#         :param str servicesModifiedAttributes: details about the service attributes after the changes
+#         """
+#         address = context.connectivity.server_address
+#         domain = context.reservation.domain
+#         token = context.connectivity.admin_auth_token
+#         return ServicesEvents(self.get_cs_session(address, domain, token), context.reservation.reservation_id).\
+#             before_my_service_removed(context, actionDetails, resourcesDetails, serviceDetails,
+#                                       oldServiceAttributeDetails, servicesRemoved, servicesAdded, modifiedServices,
+#                                       servicesModifiedAttributes)
 
 
     # </editor-fold>
@@ -331,6 +345,5 @@ class PowerService (ResourceDriverInterface):
         Destroy the driver session, this function is called everytime a driver instance is destroyed
         This is a good place to close any open sessions, finish writing to log files
         """
-        self.cs_session.Logoff
+        self.cs_session.Logoff()
         self.cs_session = None
-        pass
